Avoid warning.
[git-p4: depot-paths = "//depot/qt/main/": change = 42754]

+++ src/sql/src/odbc/qsql_odbc.cpp
@@ -222,7 +222,7 @@ QString qGetStringData( SQLHANDLE hStmt, int column, SQLINTEGER& lengthIndicator
 		    fieldVal += QString( (char*)buf );
 		    break;
 		} else {
-		    if( fieldVal.length() >= lengthIndicator ) // ### HACK - remove asap
+		    if( (int)fieldVal.length() >= lengthIndicator ) // ### HACK - remove asap
 			break;
 		    fieldVal += QString( (char*)buf );
 		}
